3rd-Person Controller Setup/Countdown Timer
Created 3rd-Person controller with a locked camera that follows the player. Camera does get "stuck" when the player walks behind an object in scene. Countdown timer functionality added to the game mode blueprint. Started to work on McGuffin pickup, but could not figure out how to get the object to appear to be held by the player after walking over the object.